Refactor upload logic out of scanner
