+++ src/views/Scanner.tsx
@@ -1,56 +1,12 @@
 import React, {useState} from 'react';
-import {
-  PermissionsAndroid,
-  Platform,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  Vibration,
-  View,
-} from 'react-native';
+import {StyleSheet, Vibration, View} from 'react-native';
 import {RNCamera} from 'react-native-camera';
-import {launchImageLibrary} from 'react-native-image-picker';
-import RNQRGenerator from 'rn-qr-generator';
 import {Props} from '../../App';
 import {CredentialHelper} from '../utils/credhelper';
 
 export const Scanner = ({navigation}: Props) => {
   const [active, setActive] = useState(true);
   const credHelper = new CredentialHelper();
-
-  async function hasAndroidPermission() {
-    const permission = PermissionsAndroid.PERMISSIONS.WRITE_EXTERNAL_STORAGE;
-
-    const hasPermission = await PermissionsAndroid.check(permission);
-    if (hasPermission) {
-      return true;
-    }
-
-    const status = await PermissionsAndroid.request(permission);
-    return status === 'granted';
-  }
-
-  async function uploadImage() {
-    if (Platform.OS === 'android' && !(await hasAndroidPermission())) {
-      return;
-    }
-
-    try {
-      setActive(false);
-      launchImageLibrary({mediaType: 'photo', base64: true}, async image => {
-        const res = await RNQRGenerator.detect({uri: image?.assets[0]?.uri});
-        if (res.values.length) {
-          for (const cred of res.values) {
-            await credHelper.storeCredential(cred);
-          }
-
-          navigation.navigate('CredentialTabs');
-        }
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  }
 
   return (
     <View style={styles.container}>
@@ -74,16 +30,6 @@ export const Scanner = ({navigation}: Props) => {
             navigation.navigate('CredentialTabs');
           }}>
           <View style={styles.window} />
-          <View
-            style={{flex: 0, flexDirection: 'row', justifyContent: 'center'}}>
-            <TouchableOpacity
-              style={styles.upload}
-              onPress={() => {
-                uploadImage();
-              }}>
-              <Text style={{fontSize: 18}}>Upload</Text>
-            </TouchableOpacity>
-          </View>
         </RNCamera>
       )}
     </View>
